refactor: 서비스에 위치한 몇몇 메서드를 도메인으로 이동
