+++ src/main/java/wooteco/subway/domain/Station.java
@@ -1,5 +1,7 @@
 package wooteco.subway.domain;
 
+import wooteco.subway.exception.DataLengthException;
+
 import java.util.Objects;
 
 public class Station {
@@ -8,12 +10,19 @@ public class Station {
     private final String name;
 
     public Station(Long id, String name) {
+        validateDataSize(name);
         this.id = id;
         this.name = name;
     }
 
     public Station(String name) {
         this(null, name);
+    }
+
+    private void validateDataSize(String name) {
+        if (name.isEmpty() || name.length() > 255) {
+            throw new DataLengthException("역 이름이 빈 값이거나 최대 범위를 초과했습니다.");
+        }
     }
 
     public Long getId() {
