+++ src/main/java/wooteco/subway/service/StationService.java
@@ -6,7 +6,6 @@ import wooteco.subway.domain.Station;
 import wooteco.subway.dto.StationRequest;
 import wooteco.subway.dto.StationResponse;
 import wooteco.subway.exception.AccessNoneDataException;
-import wooteco.subway.exception.DataLengthException;
 
 import java.util.List;
 import java.util.stream.Collectors;
@@ -21,16 +20,9 @@ public class StationService {
     }
 
     public StationResponse create(StationRequest request) {
-        validateDataSize(request.getName());
         Station station = new Station(request.getName());
         Station newStation = stationDao.insert(station);
         return StationResponse.of(newStation);
-    }
-
-    private void validateDataSize(String name) {
-        if (name.isEmpty() || name.length() > 255) {
-            throw new DataLengthException("역 이름이 빈 값이거나 최대 범위를 초과했습니다.");
-        }
     }
 
     public List<StationResponse> findAll() {
